Need to account for Tie breaking
There are a few pairs the tie because of the nature of multiplication. A*B = B*A. In this case this is not helpful and so you need to sort duplicate values differently. The multiples in this case are; (DEN -> IAH and MIA -> IAH) , (IAH->ORD, MIA->ORD), and (ATL->ORD, JFK->ORD). You choose the value where the other location is less. In the First Example, DEN(6)*IAH(3) and MIA(3)*IAH(6), you would put MIA DEN after MIA and continue. I am sure there is a more elegant way to do these calculations so this problem would not come up in the first place. Either not using multiplication or doing some preprocessing of some kind when you combine the values.